+++ test/oscope/embedded_profile_test.clj
@@ -174,8 +174,8 @@
 
 (def ^:private coordinates
   {:otel [["casselc_otel.git" "io.github.casselc/otel"]
-          ["https___github.com_casselc_otel.git/0e701ceff526d159884fadae98dcca61272ef6e0/"
-           "io.github.casselc/otel/0e701ceff526d159884fadae98dcca61272ef6e0/"]]
+          ["https___github.com_casselc_otel.git/88a63fd90e0969635dda75fbb9fe5ba2264c09d8/"
+           "io.github.casselc/otel/88a63fd90e0969635dda75fbb9fe5ba2264c09d8/"]]
    :chdb [["chucklehead-dev_jolt-chdb.git"
            "io.github.chucklehead-dev/jolt-chdb"]
           ["https___github.com_chucklehead-dev_jolt-chdb.git/19e0ecf9e9f5e2c3f24ac8758f5d6953fd021774/"